pyproject.toml
[project]
name = "bentolab"
version = "0.1.0"
description = "Python control library for Bento Lab PCR workstations"
requires-python = ">=3.11"
dependencies = [
    "bleak>=0.22.0",
    "platformdirs>=4.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.12",
]

[project.optional-dependencies]
tools = [
    "construct>=2.10.0",
    "zeroconf>=0.80.0",
    "pyshark>=0.6.0",
    "scapy>=2.5.0",
    "aiohttp>=3.9.0",
    "prompt-toolkit>=3.0.0",
]

[project.scripts]
bentolab = "bentolab.cli.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "pyright>=1.1.0",
    "complexipy>=0.8",
    "aiohttp>=3.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-m \"not hardware\""
markers = [
    "hardware: requires physical Bento Lab device",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C90"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"tools/*" = ["C901"]

[tool.pyright]
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
include = ["bentolab"]
exclude = [
    "**/__pycache__",
    "**/node_modules",
    "**/.*",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["bentolab*"]
